--- modules/@angular/upgrade/src/aot.ts
@@ -1,12 +1,0 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-export {downgradeComponent} from './aot/downgrade_component';
-export {downgradeInjectable} from './aot/downgrade_injectable';
-export {UpgradeComponent} from './aot/upgrade_component';
-export {UpgradeModule} from './aot/upgrade_module';
