+++ modules/@angular/upgrade/test/aot/integration/examples_spec.ts
@@ -10,8 +10,8 @@ import {Component, Directive, ElementRef, Injector, Input, NgModule, destroyPlat
 import {async} from '@angular/core/testing';
 import {BrowserModule} from '@angular/platform-browser';
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
-import {UpgradeComponent, UpgradeModule, downgradeComponent} from '@angular/upgrade';
 import * as angular from '@angular/upgrade/src/angular_js';
+import {UpgradeComponent, UpgradeModule, downgradeComponent} from '@angular/upgrade/static';
 
 import {bootstrap, html, multiTrim} from '../test_helpers';
