+++ modules/playground/src/bootstrap.ts
@@ -39,6 +39,7 @@
               '/packages-dist/platform-webworker-dynamic/bundles/platform-webworker-dynamic.umd.js',
           '@angular/router': '/packages-dist/router/bundles/router.umd.js',
           '@angular/upgrade': '/packages-dist/upgrade/bundles/upgrade.umd.js',
+          '@angular/upgrade/static': '/packages-dist/upgrade/bundles/upgrade-static.umd.js',
           'rxjs': '/all/playground/vendor/rxjs',
 
           // TODO(i): remove once playground apps no longer use facades directly
