+++ tools/public_api_guard/upgrade/index.d.ts
@@ -1,9 +1,3 @@
-/** @experimental */
-export declare function downgradeComponent(info: ComponentInfo): angular.IInjectable;
-
-/** @experimental */
-export declare function downgradeInjectable(token: any): (string | ((i: Injector) => any))[];
-
 /** @stable */
 export declare class UpgradeAdapter {
     constructor(ng2AppModule: Type<any>, compilerOptions?: CompilerOptions);
@@ -24,21 +18,4 @@ export declare class UpgradeAdapterRef {
     ng2ModuleRef: NgModuleRef<any>;
     dispose(): void;
     ready(fn: (upgradeAdapterRef?: UpgradeAdapterRef) => void): void;
-}
-
-/** @experimental */
-export declare class UpgradeComponent implements OnInit, OnChanges, DoCheck {
-    constructor(name: string, elementRef: ElementRef, injector: Injector);
-    ngDoCheck(): void;
-    ngOnChanges(changes: SimpleChanges): void;
-    ngOnInit(): void;
-}
-
-/** @experimental */
-export declare class UpgradeModule {
-    $injector: angular.IInjectorService;
-    injector: Injector;
-    ngZone: NgZone;
-    constructor(injector: Injector, ngZone: NgZone);
-    bootstrap(element: Element, modules?: string[], config?: angular.IAngularBootstrapConfig): void;
 }
